requirements.txt
fastapi
uvicorn
python-docx
markdown-it-py
requests
